requirements.txt
python-dotenv
pandas
numpy
requests
beautifulsoup4
lxml
tqdm
pydantic
sqlalchemy
psycopg2-binary
spacy
rapidfuzz
simhash
pydantic-settings
prefect
prophet
statsmodels
scikit-learn
matplotlib
plotly
streamlit
pyyaml
